# `Enumerable` `Find`, better performance