--- a/src/find.cr
+++ b/src/find.cr
@@ -5,17 +5,20 @@ module Find
     next Find::Skip
   end
 
+  # Skips dangling symlinks for compatibility with ruby
   def self.find(*paths)
+    super(*paths) do |path|
+      exists = File.exists?(path) rescue false
+      yield(path) if exists
+    end
+  end
+  
+  def self.find_all(*paths)
     paths.each do |path|
       search_path = [path]
       while !search_path.empty?
         file = search_path.shift
-
-        begin
-          next if file.nil? || !File.exists?(file)
-        rescue
-          next
-        end
+        next if file.nil?
 
         skip = yield file.dup
         next if skip == Find::Skip