`Find` is `Enumerable` and yields `{Path, File::Info}`
Use `push`/`pop` instead of `shift`/`unshift`
`follow_symlinks: false` to avoid infinite loops

### Benchmarks

Before
```
 Find.find double stat   1.03k (970.82µs) (± 3.70%)  53.2kB/op        fastest
```

After
```
 Find.find double stat   2.72M (367.65ns) (± 5.59%)   224B/op   4.96× slower
  Find.new double stat  12.63M ( 79.19ns) (± 7.76%)  96.0B/op   1.07× slower
  Find.new single stat  13.50M ( 74.09ns) (± 9.76%)  96.0B/op        fastest
```

`Find.new double stat` doesn't perform `File.exists` explaining the speed boost from `Find.find double stat`.
All `double stat` runs are for comparison with the original code.

Benchmark:
```crystal
require "benchmark"
require "./src/find"

stats = Hash(Symbol, Int32).new { |h, k| h[k] = 0 }

Benchmark.ips do |bm|
  bm.report "Find.find double stat" do
    stats.clear
    Find.find(".") do |path|
      if info = File.info? path, follow_symlinks: false
        if info.file?
          stats[:file] += 1
        elsif info.directory?
          stats[:dir] += 1
        end
      end
    end
  end

  bm.report "Find.new double stat" do
    stats.clear
    Find.new(["."]).each do |path, _|
      if info = File.info? path, follow_symlinks: false
        if info.file?
          stats[:file] += 1
        elsif info.directory?
          stats[:dir] += 1
        end
      end
    end
  end

  bm.report "Find.new single stat" do
    stats.clear
    Find.new(["."]).each do |path, info|
      if info.file?
        stats[:file] += 1
      elsif info.directory?
        stats[:dir] += 1
      end
    end
  end
end
```