@didactic-drunk 2022-04-05
### `Enumerable` uses:

One line `du`: 
```crystal
p Find.new(ARGV).select { |_, info| info.file? }.map { _, info.size }.sum
```

Parallel file processing with [concurrent.cr](https://github.com/didactic-drunk/concurrent.cr)
```crystal
Find.new(...).parallel.select { |path, info| 
  info.file? && path.to_s =~ /\.json/
}.run { |path, _| }
  File.open(path) do |rfile|
    File.atomic_rewrite("#{path.basename}.newext" do |wfile|
      ...
    end
  end
}.wait
```