@didactic-drunk 2022-04-07
Pondering this more, I think `File.exists?` should be removed
1) It's utility is dubious.  Either a dangling symlink was followed or there is a race condition which can NOT be solved with `File.exists?`
2) Performance penalty especially on remote file systems